portia_demo/backup/sql_mcp_server/pyproject.toml
[tool.poetry]
name = "sql-mcp-server"
version = "0.1.0"
description = "MCP server for SQL operations"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
modelcontextprotocol = "^0.12.0"
sqlalchemy = "^2.0.28"
pydantic = "^2.5.2"
sqlite3worker = "^1.1.6"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 